Split functions with linux and windows versions
